refactor transitions
