+++ test/helpers/spectron/network.ts
@@ -1,10 +1,11 @@
-import { focusWorker, TExecutionContext } from './index';
+import { TExecutionContext } from './index';
+import { focusWindow } from '../modules/core';
 
 /**
  * install the fetch-mock lib
  */
 export async function installFetchMock(t: TExecutionContext) {
-  await focusWorker(t);
+  await focusWindow('worker');
   await t.context.app.webContents.executeJavaScript(`
     window.fetchMock = require('fetch-mock');
     0; // Prevent returning a value that cannot be serialized
@@ -15,7 +16,7 @@ export async function installFetchMock(t: TExecutionContext) {
  * mock fetch requests
  */
 export async function fetchMock(t: TExecutionContext, regExp: RegExp, code: number) {
-  await focusWorker(t);
+  await focusWindow('worker');
   await t.context.app.webContents.executeJavaScript(`
     fetchMock.mock(${regExp.toString()}, ${code});
     0; // Prevent returning a value that cannot be serialized
@@ -26,7 +27,7 @@ export async function fetchMock(t: TExecutionContext, regExp: RegExp, code: numb
  * reset all mocks
  */
 export async function resetFetchMock(t: TExecutionContext) {
-  await focusWorker(t);
+  await focusWindow('worker');
   await t.context.app.webContents.executeJavaScript(`
     fetchMock.reset();
     0; // Prevent returning a value that cannot be serialized
