+++ test/helpers/widget-helpers.ts
@@ -1,8 +1,9 @@
 import { sleep } from './sleep';
-import { focusChild, TExecutionContext } from './spectron';
+import { TExecutionContext } from './spectron';
 import { WidgetsService, WidgetType } from '../../app/services/widgets';
 import { SourcesService } from '../../app/services/sources';
 import { getClient } from './api-client';
+import {focusChild} from "./modules/core";
 
 export async function waitForWidgetSettingsSync(t: TExecutionContext) {
   await sleep(2000);
@@ -42,6 +43,6 @@ export async function addWidget(t: TExecutionContext, type: EWidgetType, name: s
 
   const widget = widgetService.createWidget((type as unknown) as WidgetType, name);
   sourcesService.showSourceProperties(widget.sourceId);
-  await focusChild(t);
+  await focusChild();
   await (await t.context.app.client.$('button=Widget Editor')).waitForDisplayed(); // wait for loading
 }
