+++ test/regular/selective-recording.ts
@@ -1,10 +1,11 @@
 import { readdir } from 'fs-extra';
-import { focusMain, test, useSpectron } from '../helpers/spectron';
+import { test, useSpectron } from '../helpers/spectron';
 import { addSource } from '../helpers/modules/sources';
 import {
   setOutputResolution,
   setTemporaryRecordingPath,
 } from '../helpers/modules/settings/settings';
+import { focusMain } from '../helpers/modules/core';
 
 useSpectron();
 
@@ -21,7 +22,7 @@ test('Selective Recording', async t => {
   await addSource(sourceType, sourceName);
 
   // Toggle selective recording
-  await focusMain(t);
+  await focusMain();
   await (await client.$('.studio-controls-top .icon-smart-record')).click();
 
   // Check that selective recording icon is active
