+++ test/regular/settings/audio.ts
@@ -1,6 +1,7 @@
-import { focusMain, test, useSpectron } from '../../helpers/spectron';
+import { test, useSpectron } from '../../helpers/spectron';
 import { assertOptions } from '../../helpers/spectron/assertions';
 import { showSettingsWindow } from '../../helpers/modules/settings/settings';
+import {focusMain} from "../../helpers/modules/core";
 
 useSpectron();
 
@@ -33,7 +34,7 @@ test('Populates audio settings', async t => {
 
   // Test that we're displaying mixer settings in the footer
   await (await app.client.$('button=Done')).click();
-  await focusMain(t);
+  await focusMain();
   t.true(await (await app.client.$('.source-name=Desktop Audio')).isExisting());
   t.true(await (await app.client.$('.source-name*=Mic')).isExisting());
   t.true(await (await app.client.$('.volmeter-container')).isExisting());
