+++ test/regular/transitions.ts
@@ -1,10 +1,11 @@
-import { useSpectron, focusChild, focusMain, test } from '../helpers/spectron';
+import { useSpectron, test } from '../helpers/spectron';
 import { clickSceneTransitions, addScene } from '../helpers/modules/scenes';
 import { getFormInput } from '../helpers/spectron/forms';
 import { dismissModal } from '../helpers/spectron/modals';
 import { FormMonkey } from '../helpers/form-monkey';
+import { click, clickButton, focusChild, focusMain } from '../helpers/modules/core';
 
-useSpectron();
+useSpectron({ implicitTimeout: 0 });
 
 test('Changing transition options', async t => {
   const app = t.context.app;
@@ -14,9 +15,9 @@ test('Changing transition options', async t => {
   // We need at least 2 scenes to edit transitions
   await addScene('Other Scene');
 
-  await focusMain(t);
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
   await (await app.client.$('.icon-edit')).click();
   const form = new FormMonkey(t);
   await form.fillByTitles({
@@ -25,18 +26,22 @@ test('Changing transition options', async t => {
   });
 
   await dismissModal(t);
-  await (await t.context.app.client.$('button=Done')).click();
-  await focusMain(t);
+  await clickButton('Done');
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
 
-  await (await app.client.$('.icon-edit')).click();
+  await click('.icon-edit');
+
+  console.log('includesByTitles');
+
   t.true(
     await form.includesByTitles({
       Type: transitionType,
       Duration: transitionDuration,
     }),
   );
+  t.pass();
 });
 
 test('Adding and removing transitions', async t => {
@@ -45,9 +50,9 @@ test('Adding and removing transitions', async t => {
   // We need at least 2 scenes to edit transitions
   await addScene('Other Scene');
 
-  await focusMain(t);
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
   await (await app.client.$('button=Add Transition')).click();
   await dismissModal(t);
   await (await app.client.$('.icon-trash')).click();
@@ -65,9 +70,9 @@ test('Changing connections', async t => {
   // We need at least 2 scenes to edit transitions
   await addScene('Other Scene');
 
-  await focusMain(t);
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
   await (await app.client.$('button=Add Transition')).click();
   await dismissModal(t);
   await (await app.client.$('button=Connections')).click();
@@ -79,9 +84,9 @@ test('Changing connections', async t => {
     'Ending Scene': connectionEnd,
   });
   await (await t.context.app.client.$('button=Done')).click();
-  await focusMain(t);
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
 
   await (await app.client.$('button=Connections')).click();
   await (await app.client.$('.icon-edit')).click();
@@ -101,9 +106,9 @@ test('Showing redudant connection warning', async t => {
   // We need at least 2 scenes to edit transitions
   await addScene('Other Scene');
 
-  await focusMain(t);
+  await focusMain();
   await clickSceneTransitions();
-  await focusChild(t);
+  await focusChild();
   await (await app.client.$('button=Add Transition')).click();
   await dismissModal(t);
   await (await app.client.$('button=Connections')).click();
