+++ test/screentest/tests/settings.ts
@@ -1,9 +1,10 @@
-import { useSpectron, test, focusChild, TExecutionContext, closeWindow } from '../../helpers/spectron';
+import { useSpectron, test, TExecutionContext } from '../../helpers/spectron';
 import { getClient } from '../../helpers/api-client';
 import { makeScreenshots, useScreentest } from '../screenshoter';
 import { SettingsService } from '../../../app/services/settings';
 import { logIn, logOut } from '../../helpers/spectron/user';
 import { sleep } from '../../helpers/sleep';
+import {closeWindow, focusChild} from "../../helpers/modules/core";
 
 
 useSpectron({ restartAppAfterEachTest: false });
@@ -13,7 +14,7 @@ test('Settings General', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings();
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -21,7 +22,7 @@ test('Settings Stream Offline', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Stream');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -30,7 +31,7 @@ test('Settings Stream Twitch', async (t: TExecutionContext) => {
   const settingsService = client.getResource<SettingsService>('SettingsService');
   await logIn(t, 'twitch');
   settingsService.showSettings('Stream');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -39,7 +40,7 @@ test('Settings Stream Youtube', async (t: TExecutionContext) => {
   const settingsService = client.getResource<SettingsService>('SettingsService');
   await logIn(t, 'youtube');
   settingsService.showSettings('Stream');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -48,7 +49,7 @@ test('Settings Stream Facebook', async (t: TExecutionContext) => {
   const settingsService = client.getResource<SettingsService>('SettingsService');
   await logIn(t, 'facebook');
   settingsService.showSettings('Stream');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -56,7 +57,7 @@ test('Settings Output', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Output');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -64,7 +65,7 @@ test('Settings Video', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Video');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -72,7 +73,7 @@ test('Settings Hotkeys', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Hotkeys');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -80,7 +81,7 @@ test('Settings Scene Collections', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Scene Collections');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -88,7 +89,7 @@ test('Settings Notifications', async t => {
   const client = await getClient();
   const settingsService = client.getResource<SettingsService>('SettingsService');
   settingsService.showSettings('Notifications');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -97,7 +98,7 @@ test('Settings Appearance', async t => {
   const settingsService = client.getResource<SettingsService>('SettingsService');
   await sleep(1000);
   settingsService.showSettings('Appearance');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -107,14 +108,14 @@ test('Settings Game Overlay', async (t: TExecutionContext) => {
 
   // take offline screenshot
   settingsService.showSettings('Game Overlay');
-  await focusChild(t);
+  await focusChild();
   await makeScreenshots(t, 'Offline');
-  await closeWindow(t);
+  await closeWindow('child');
 
   // take online screenshot
   await logIn(t);
   settingsService.showSettings('Game Overlay');
-  await focusChild(t);
+  await focusChild();
   await (await t.context.app.client.$('[data-type="toggle"]')).click(); // enable overlays
   await makeScreenshots(t, 'Online');
   await logOut(t);
