+++ test/screentest/tests/sources.ts
@@ -1,8 +1,9 @@
-import { useSpectron, test, focusChild } from '../../helpers/spectron';
+import { useSpectron, test } from '../../helpers/spectron';
 import { getClient } from '../../helpers/api-client';
 import { ISourcesServiceApi } from '../../../app/services/sources/sources-api';
-import { makeScreenshots, useScreentest } from '../screenshoter';
+import { useScreentest } from '../screenshoter';
 import { ScenesService } from 'services/api/external-api/scenes';
+import { focusChild } from '../../helpers/modules/core';
 
 useSpectron({ restartAppAfterEachTest: false });
 useScreentest();
@@ -11,7 +12,7 @@ test('Sources showcase window', async t => {
   const client = await getClient();
   const sourcesService = client.getResource<ISourcesServiceApi>('SourcesService');
   sourcesService.showShowcase();
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -19,7 +20,7 @@ test('AddSource window', async t => {
   const client = await getClient();
   const sourcesService = client.getResource<ISourcesServiceApi>('SourcesService');
   sourcesService.showAddSource('color_source');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
 
@@ -29,7 +30,6 @@ test('AddSource window with suggestions', async t => {
   const scenesService = client.getResource<ScenesService>('ScenesService');
   scenesService.activeScene.createAndAddSource('MySource', 'color_source');
   sourcesService.showAddSource('color_source');
-  await focusChild(t);
+  await focusChild();
   t.pass();
 });
-
