+++ test/screentest/tests/transitions.ts
@@ -1,8 +1,9 @@
-import { useSpectron, test, focusChild, TExecutionContext } from '../../helpers/spectron';
+import { useSpectron, test, TExecutionContext } from '../../helpers/spectron';
 import { getClient } from '../../helpers/api-client';
 import { makeScreenshots, useScreentest } from '../screenshoter';
 import { ScenesService } from 'services/api/external-api/scenes';
 import { TransitionsService } from 'services/transitions';
+import { focusChild } from '../../helpers/modules/core';
 
 useSpectron({ restartAppAfterEachTest: false });
 useScreentest();
@@ -13,7 +14,7 @@ test('Transitions', async (t: TExecutionContext) => {
   const transitionService = client.getResource<TransitionsService>('TransitionsService');
 
   transitionService.showSceneTransitions();
-  await focusChild(t);
+  await focusChild();
   await makeScreenshots(t, 'No scenes');
   scenesService.createScene('New Scene');
   await makeScreenshots(t, '1 scene');
