+++ test/screentest/tests/widgets/goals.ts
@@ -1,8 +1,9 @@
-import { TExecutionContext, test, useSpectron, closeWindow } from '../../../helpers/spectron/index';
+import { TExecutionContext, test, useSpectron } from '../../../helpers/spectron/index';
 import { logIn, logOut } from '../../../helpers/spectron/user';
 import { makeScreenshots, useScreentest } from '../../screenshoter';
 import { FormMonkey } from '../../../helpers/form-monkey';
 import { addWidget, EWidgetType, waitForWidgetSettingsSync } from '../../../helpers/widget-helpers';
+import { closeWindow } from '../../../helpers/modules/core';
 
 useSpectron({ restartAppAfterEachTest: false });
 useScreentest();
@@ -52,7 +53,7 @@ function testGoal(goalType: string, widgetType: EWidgetType) {
     `);
     await makeScreenshots(t, 'Created Goal');
 
-    await closeWindow(t);
+    await closeWindow('child');
     await logOut(t);
   });
 
@@ -80,7 +81,7 @@ function testGoal(goalType: string, widgetType: EWidgetType) {
 
     await makeScreenshots(t, 'Settings');
 
-    await closeWindow(t);
+    await closeWindow('child');
     await logOut(t);
     t.pass();
   });
